update-price-table/requirements.txt

boto3
bs4
python-dotenv
requests
psycopg2-binary
